fix: enable unicode-case feature of regex
to fix failures in binding generation.

We use case-inensistive regex patterns while calling allowlist_file() at some
places. This does not work without the unicode-case feature. In the past it
used to be enabled by tracing-subscriber which served our allowlist-file()
use-case as well by accident. But the new version of tracing-subscriber
has removed the dependency on the regex crate thereboy breaking our
bindgen code.

In this fix we have added an explicit depenency on regex ourselves with
said feature enabled.